feat: Add DeployMoodNft script and test
This commit adds the DeployMoodNft script and its corresponding test. The script allows for the deployment of the MoodNft contract, while the test verifies the functionality of converting an SVG image to a URI. This is an important addition to the project as it enables the creation and tracking of mood-based NFTs.