refactor: Update file paths in foundry.toml and DeployMoodNft.s.sol
This commit updates the file paths in the `foundry.toml` and `DeployMoodNft.s.sol` files. The path for the images directory in `foundry.toml` has been changed from "./images/" to "./img/". In `DeployMoodNft.s.sol`, the script now reads the SVG files from the "./img/dynamicNft/" directory instead of "./img/". This change ensures that the correct files are accessed and used in the deployment process.