feat: Add BagelToken contract
Add the BagelToken contract which is an ERC20 token contract named "Bagel" with the symbol "BAGEL". The contract extends the ERC20 and Ownable contracts from the OpenZeppelin library. It includes a constructor to set the token name and symbol, and a mint function that can only be called by the contract owner to mint new tokens.